data/credentials/fixture/pyproject.toml
[tool.poetry]
name = "pdf_excel_parser"
version = "0.1.0"
description = "Project for custom xlsx file enrichment"
authors = ["Dmytro Kovtun"]

[tool.poetry.dependencies]
python = ">3.8.1"
google-api-python-client = "^2.66.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
